refactor: redistribute logic in App and Pagination components
